// packages/@zenithui/toast/src/lib/types.ts
export type ToastType =
  | 'success'
  | 'info'
  | 'error'
  | 'warning'
  | 'loading'
  | 'promise';

export type Theme = 'auto' | 'light' | 'dark';

export type ToastPosition =
  | 'top-left'
  | 'top-right'
  | 'bottom-left'
  | 'bottom-right'
  | 'top-center'
  | 'bottom-center';

export type ToastAnimation = 'slide' | 'fade' | 'enter-with-icon';

export type ClassNames = {
  /**
   * The class name of the toast container.
   */
  className: string;
  /**
   * The class name of the toast icon.
   */
  icon: string;
  /**
   * The class name of the toast message.
   */
  title: string;
  /**
   * The class name of the toast description.
   */
  description: string;
  /**
   * The class name of the toast close button.
   */
  closeButton: string;
  /**
   * The class name of the toast action button.
   */
  actionButton: string;
  /**
   * The class name of the toast cancel button.
   */
  cancelButton: string;
};

export type ToastOptions = {
  /**
   * Whether to use rich colors for the toast.
   * @type {boolean}
   * @default false
   */
  richColors?: boolean;
  /**
   * The duration of the toast to display.
   * @type {number}
   * @default 5000
   */
  duration?: number;
  /**
   * Whether to enable auto dismiss for the toast.
   * @type {boolean}
   * @default true
   */
  disableAutoDismiss?: boolean;
  /**
   * The animation of the toast.
   * @type {string}
   * @default "fade"
   */
  animation?: ToastAnimation;
  /**
   * Whether to show the close button for the toast.
   * @type {boolean}
   * @default false
   */
  showCloseButton?: boolean;
  /**
   * The position of the toast.
   * @type {ToastPosition}
   * @default "top-right"
   * @description This is used to position the toast.
   */
  position?: ToastPosition;
  /**
   * The title of the toast.
   * @type {string}
   * @description This is used to display the title of the toast.
   */
  title?: string;
  /**
   * The description of the toast.
   * @type {string}
   * @description This is used to display the description of the toast.
   */
  description?: string;
  /**
   * The classNames to customise the toast.
   * @type {Partial<ClassNames>}
   */
  classNames?: Partial<ClassNames> | string;
  /**
   * The callback function to execute when the action button is clicked.
   * @type {React.MouseEventHandler<HTMLButtonElement>}
   */
  onAction?: React.MouseEventHandler<HTMLButtonElement>;
  /**
   * The callback function to execute when the cancel button is clicked.
   * @type {React.MouseEventHandler<HTMLButtonElement>}
   */
  onCancel?: React.MouseEventHandler<HTMLButtonElement>;
  /**
   * The callback function to execute when the toast is closed.
   * @type {React.MouseEventHandler<HTMLButtonElement>}
   */
  onClose?: React.MouseEventHandler<HTMLButtonElement>;
  /**
   * The custom action button.
   * @type {React.FC<ButtonProps>}
   */
  action?: React.FC<ButtonProps>;
  /**
   * The custom cancel button.
   * @type {React.FC<ButtonProps>}
   */
  cancel?: React.FC<ButtonProps>;
  /**
   * The custom close button.
   * @type {React.FC<ButtonProps>}
   */
  close?: React.FC<ButtonProps>;
  /**
   * The icon to display in the toast.
   * @type {React.ReactNode}
   */
  icon?: React.ReactNode;
};
// ...
export interface ButtonProps extends React.HTMLAttributes<HTMLButtonElement> {
  btntype?: 'action' | 'cancel' | 'close';
}
// ...
export interface PromiseToast {
  loading?: string;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  success?: string | ((data: any) => string);
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  error?: string | ((data: any) => string);
}